Possibility to make a dropdown from a single select
